+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Add list of tenants in alloy config methods's parameters.
+- Fine-tune alloy-metrics resource usage configuration to avoid causing issues for customer workload and cluster tests.
 
 ## [0.20.0] - 2025-03-18