+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.21.0] - 2025-03-24
+
 ### Added
 
 - Add multi-tenancy support to alloy remote write by creating a custom remote-write section per tenant defined in Grafana Organization CRs.
@@ -363,7 +365,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initialize project and create heartbeat for the installation.
 
-[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.20.0...HEAD
+[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.21.0...HEAD
+[0.21.0]: https://github.com/giantswarm/observability-operator/compare/v0.20.0...v0.21.0
 [0.20.0]: https://github.com/giantswarm/observability-operator/compare/v0.19.4...v0.20.0
 [0.19.4]: https://github.com/giantswarm/observability-operator/compare/v0.19.3...v0.19.4
 [0.19.3]: https://github.com/giantswarm/observability-operator/compare/v0.19.2...v0.19.3